+++ NEWS.md
@@ -4,7 +4,7 @@
     - now takes argument n to specify number of classes
     - style
 - Removed Roboto font and `util_import_roboto`
-- Removed `util_plo_grey`
+- Removed `util_plot_grey`
 - Renamed:
     - `util_plot` to `show_landscape`
 - new function `util_writeESRI` that produces a replica of esris ascii file format